refactor : guide 내용들 파일 분리해서 적용
