+++ src/components/screens/ShootingGuide.tsx
@@ -6,25 +6,7 @@ import { font, os } from "@/style/font.ts";
 import CameraSvg from "@assets/svgs/camera.svg";
 import { requestCameraPermission } from "@/utils/permission";
 import { useNavigation } from "@react-navigation/native";
-
-const shootingGuideData= [
-  {
-    title: '네모칸 안에 알약이 보이도록 촬영해주세요',
-    description: '알약이 잘리면 정확한 결과가 안나올 수 있어요',
-    mainImage: require('@/assets/images/search_step1_main.png'),
-    subImage: require('@/assets/images/search_step1_sub.png') },
-  {
-    title: '알약에 글자가 선명히 보이도록 촬영해주세요',
-    description: '알약의 글자가 잘리거나 흐리게 촬영하면 정확한 결과가 안나올 수 있어요',
-    mainImage: require('@/assets/images/search_step2_main.png'),
-    subImage: require('@/assets/images/search_step2_sub.png')
-  },
-  {
-    title: '알약은 하나씩만 찍어주세요.',
-    description: '붙어있거나 겹쳐있으면 정확한 결과가 안나올 수 있어요',
-    mainImage: require('@assets/images/search_step3_main.png'),
-    subImage: require('@/assets/images/search_step3_sub.png') },
-];
+import { shootingGuideData } from "@/constants/guide";
 
 const { width } = Dimensions.get('window');
 
@@ -181,18 +163,20 @@ const styles = StyleSheet.create({
     position: "relative",
     alignItems: "center",
     gap: 10,
-    // top: -10
+    fontFamily: os.font(500, 500),
   },
   mainDescription: {
     textAlign: "center",
     fontWeight: "700",
-    fontSize: font(16)
+    fontSize: font(16),
+    fontFamily: os.font(500, 500),
   },
   subDescription: {
     textAlign: "center",
     fontWeight: "700",
     fontSize: font(14),
-    color: "#656565"
+    color: "#656565",
+    fontFamily: os.font(500, 500),
   },
 
   button: {
