+++ src/constants/guide.ts
@@ -1,0 +1,21 @@
+const shootingGuideData= [
+  {
+    title: '네모칸 안에 알약이 보이도록 촬영해주세요',
+    description: '알약이 잘리면 정확한 결과가 안나올 수 있어요',
+    mainImage: require('@/assets/images/search_step1_main.png'),
+    subImage: require('@/assets/images/search_step1_sub.png') },
+  {
+    title: '알약에 글자가 선명히 보이도록 촬영해주세요',
+    description: '알약의 글자가 잘리거나 흐리게 촬영하면 정확한 결과가 안나올 수 있어요',
+    mainImage: require('@/assets/images/search_step2_main.png'),
+    subImage: require('@/assets/images/search_step2_sub.png')
+  },
+  {
+    title: '알약은 하나씩만 찍어주세요.',
+    description: '붙어있거나 겹쳐있으면 정확한 결과가 안나올 수 있어요',
+    mainImage: require('@assets/images/search_step3_main.png'),
+    subImage: require('@/assets/images/search_step3_sub.png') },
+];
+
+
+export { shootingGuideData };
